# dashboard/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "XDC SkyNet API",
    "description": "The definitive dashboard and API platform for XDC Network owners and operators",
    "version": "1.0.0",
    "contact": {
      "name": "XDC Network",
      "url": "https://github.com/AnilChinchawale/XDCSkyNet"
    }
  },
  "servers": [
    {
      "url": "https://net.xdc.network/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3005/api/v1",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/alerts": {
      "get": {
        "summary": "List alerts/incidents",
        "description": "Get a list of active or historical alerts",
        "tags": ["Alerts"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/IncidentStatus"
          },
          {
            "$ref": "#/components/parameters/Severity"
          }
        ],
        "responses": {
          "200": {
            "description": "List of alerts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create alert",
        "description": "Create a new alert or incident",
        "tags": ["Alerts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Alert"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alert created"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alert": {
        "type": "object",
        "required": ["type", "severity", "title"],
        "properties": {
          "nodeId": { "type": "string", "format": "uuid" },
          "type": { "type": "string", "maxLength": 50 },
          "severity": { "type": "string", "enum": ["critical", "warning", "info"] },
          "title": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "maxLength": 2000 },
          "suggestedFix": { "type": "string", "maxLength": 2000 }
        }
      },
      "AlertListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "array",
            "items": {
              "allOf": [
                { "$ref": "#/components/schemas/Alert" },
                {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "status": { "type": "string", "enum": ["active", "acknowledged", "resolved"] },
                    "detected_at": { "type": "string", "format": "date-time" },
                    "resolved_at": { "type": "string", "format": "date-time" }
                  }
                }
              ]
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": { "type": "string" },
        "description": "Pagination cursor"
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "default": 50, "maximum": 100 },
        "description": "Number of items per page"
      },
...
      "IncidentStatus": {
        "name": "status",
        "in": "query",
        "schema": { "type": "string", "enum": ["active", "acknowledged", "resolved"] },
        "description": "Filter by incident status"
      },
      "Severity": {
        "name": "severity",
        "in": "query",
        "schema": { "type": "string", "enum": ["critical", "warning", "info"] },
        "description": "Filter by severity"
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Use your API key as the bearer token."
      }
    }
  }
}